// workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/enums/Categories.ts
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export enum MarketplaceCategories {
	MARKETPLACE_APP_CATEGORY = 'marketplace-app-category',
	MARKETPLACE_APP_TAGS = 'marketplace-app-tags',
	MARKETPLACE_CATEGORY = 'marketplace-category',
	MARKETPLACE_LIFERAY_PLATFORM_OFFERING = 'marketplace-liferay-platform-offering',
	MARKETPLACE_LIFERAY_VERSION = 'marketplace-liferay-version',
	MARKETPLACE_PRODUCT_TYPE = 'marketplace-product-type',
	MARKETPLACE_SOLUTION_CATEGORY = 'marketplace-solution-category',
	MARKETPLACE_SOLUTION_EDTION = 'marketplace-solution-edition',
	MARKETPLACE_SOLUTION_TAGS = 'marketplace-solution-tags',
}

export enum MarketplaceCategory {
	BATCH = 'Batch',
	CHECKOUT = 'Checkout',
	FRAGMENTS = 'Fragments',
	OBJECT_ACTION = 'Object action',
	OTHER = 'Other',
	PAYMENT_METHODS = 'Payment methods',
	PROMPT = 'Prompt',
	SITE_INITIALIZER = 'Site Initializer',
	THEME = 'Theme',
	WORKFLOW_ACTION = 'Workflow Action',
}